requirements.txt
pandas
requests
python-dotenv
pytz
apache-airflow==2.9.0
apache-airflow-providers-postgres
